ci: fix clippy errors triggered under -Dwarnings
CI runs clippy/tests under RUSTFLAGS=-Dwarnings (set by setup-rust-toolchain),
which surfaces lints that bare `cargo clippy` doesn't. Fixes the failures from
run 25109911630:

- cast_possible_truncation: prefer try_from / Pos0 deref / saturating ops
- arithmetic_side_effects: saturating_sub/add in iterator + MD encoder
- doc_markdown: backtick OwnedBamRecord, push_raw, next_pos, MD return tuple
- field_reassign_with_default: struct literal for AlignedPairsOptions
- useless_conversion / unnecessary_cast cleanups
- example: CustomizeRecordStore::keep_record → filter (trait method rename)

Add a "Pre-push CI parity" section to CLAUDE.md with the exact local commands
and the recurring lint patterns to watch for.

+++ crates/seqair/src/bam/record_store.rs
@@ -1752,19 +1752,15 @@ mod tests {
                 reason = "bases.len() bounded ≤ 16 by strategy; fits in u32"
             )]
             let len = input.bases.len() as u32;
-            OwnedBamRecord::builder(
-                0,
-                Some(Pos0::new(u32::from(input.pos)).unwrap()),
-                input.qname.clone(),
-            )
-            .flags(BamFlags::empty())
-            .mapq(input.mapq)
-            .cigar(vec![CigarOp::new(cigar::CigarOpType::Match, len)])
-            .seq(input.bases.clone())
-            .qual(input.quals.iter().copied().map(BaseQuality::from_byte).collect())
-            .aux(AuxData::from_bytes(input.aux.clone()))
-            .build()
-            .expect("synthetic OwnedBamRecord must build")
+            OwnedBamRecord::builder(0, Some(Pos0::new(input.pos).unwrap()), input.qname.clone())
+                .flags(BamFlags::empty())
+                .mapq(input.mapq)
+                .cigar(vec![CigarOp::new(cigar::CigarOpType::Match, len)])
+                .seq(input.bases.clone())
+                .qual(input.quals.iter().copied().map(BaseQuality::from_byte).collect())
+                .aux(AuxData::from_bytes(input.aux.clone()))
+                .build()
+                .expect("synthetic OwnedBamRecord must build")
         }
 
         proptest! {
